Retyping topics to be able to work live on vehicles with the new can_to_ros framework. The speed planner information is still missing, unfortunately

+++ onnx2ros/src/NetworkReader.cpp
@@ -126,7 +126,7 @@ void PromptReader::callback_accel(const std_msgs::Float64& accel_msg) {
   prev_accels.pop_back();
 }
 
-void PromptReader::callback_minicar(const std_msgs::Float64& minicar_msg) {
+void PromptReader::callback_minicar(const std_msgs::Int16& minicar_msg) {
   state_minicar = minicar_msg;
 }
